docker-compose.yml: Wait for the database on first init
The "depends_on" option of Compose waits for the container to be
available (most likely reachable by ip address) and does not
wait for the service in that container to be ready before
continuing.

That could cause a problem since the nyaa container couldn't create
the required database tables.

As a solution, use an external script (wait-for-it) which will wait
until the database is readily accessible from it's TCP port.

A default timeout value of 30 seconds should be enough for the
first database initialization.

Credits: https://github.com/vishnubob/wait-for-it